Taintmem (#37)
* code cleanup and reformat for readability

* reformatting for readability

* reformat for readability, add comments

* reformat for readability, add comments

* reformat for readability, add comments

* reformat for readability, add comments

* refactor, reformat, and add comments

* refactor, reformat, and add comments

* reformat for ease of understanding; add comments

* reformat for ease of understanding; add comments

* reformat for ease of understanding; add comments

* reformat for ease of understanding; add comments

* reformat for ease of understanding; add comments

* reformat for ease of understanding; add comments

* reformat for ease of understanding; add comments

* tutorials

* tutorials

* tutorials

* tutorials

* added text to the tutorial

* added text to the tutorial

* first attempt at a taint propagation tool

* edits

* edits

* edits

* edits

* edits

* added option to selectively dump taint info

* added text to README

* bug fix for valgrind error: cond jump on uninitialized memory

* addd text

* addd text

* added text

* removed -fpic from CFLAGS

* slightly more efficient code for checking when to add/dump taint

+++ memtaint/main.h
@@ -23,7 +23,7 @@
 typedef struct taint_loc {
   uint64_t ins_num;  
   uint64_t start;
-  int sz;
+  uint64_t sz;
   struct taint_loc *next;
 } Taint_Loc;
